+++ sync-for-payables/src/sdk/models/operations/updatecompany.ts
@@ -4,6 +4,9 @@
 
 import * as z from "zod";
 import { remap as remap$ } from "../../../lib/primitives.js";
+import { safeParse } from "../../../lib/schemas.js";
+import { Result as SafeParseResult } from "../../types/fp.js";
+import { SDKValidationError } from "../errors/sdkvalidationerror.js";
 import * as shared from "../shared/index.js";
 
 export type UpdateCompanyRequest = {
@@ -59,4 +62,22 @@ export namespace UpdateCompanyRequest$ {
   export const outboundSchema = UpdateCompanyRequest$outboundSchema;
   /** @deprecated use `UpdateCompanyRequest$Outbound` instead. */
   export type Outbound = UpdateCompanyRequest$Outbound;
+}
+
+export function updateCompanyRequestToJSON(
+  updateCompanyRequest: UpdateCompanyRequest,
+): string {
+  return JSON.stringify(
+    UpdateCompanyRequest$outboundSchema.parse(updateCompanyRequest),
+  );
+}
+
+export function updateCompanyRequestFromJSON(
+  jsonString: string,
+): SafeParseResult<UpdateCompanyRequest, SDKValidationError> {
+  return safeParse(
+    jsonString,
+    (x) => UpdateCompanyRequest$inboundSchema.parse(JSON.parse(x)),
+    `Failed to parse 'UpdateCompanyRequest' from JSON`,
+  );
 }