+++ sync-for-payables/src/sdk/models/shared/attachment.ts
@@ -3,6 +3,9 @@
  */
 
 import * as z from "zod";
+import { safeParse } from "../../../lib/schemas.js";
+import { Result as SafeParseResult } from "../../types/fp.js";
+import { SDKValidationError } from "../errors/sdkvalidationerror.js";
 
 /**
  *   The Codat API supports pulling and pushing of file attachments for invoices, bills, direct costs, and direct incomes.
@@ -126,4 +129,18 @@ export namespace Attachment$ {
   export const outboundSchema = Attachment$outboundSchema;
   /** @deprecated use `Attachment$Outbound` instead. */
   export type Outbound = Attachment$Outbound;
+}
+
+export function attachmentToJSON(attachment: Attachment): string {
+  return JSON.stringify(Attachment$outboundSchema.parse(attachment));
+}
+
+export function attachmentFromJSON(
+  jsonString: string,
+): SafeParseResult<Attachment, SDKValidationError> {
+  return safeParse(
+    jsonString,
+    (x) => Attachment$inboundSchema.parse(JSON.parse(x)),
+    `Failed to parse 'Attachment' from JSON`,
+  );
 }